Bug#36846364 Set sanitizer environment variables for unit tests
Set ASAN_OPTIONS, LSAN_OPTIONS and UBSAN_OPTIONS for unit tests.
In particular: make UBSAN fail-fast.
The environment will be set if tests are executed using 'ctest'.

Fix four UBSAN warnings
mysys_my_time-t.cc:474:3: runtime error: signed integer overflow

hypergrap_optimizer-t
make_join_hypergraph.cc:3310:59: runtime error: reference binding to null pointer of type 'const CompanionSet'

[ RUN      ] CacheStorageTest.BasicStoreAndGetCollation
sql/dd/impl/types/collation_impl.h:47:7: runtime error:
load of value 9, which is not a valid value for type 'enum_pad_attribute'

[ RUN      ] OptRangeTest.CloneSpatialKey
sql/range_optimizer/tree.cc:353:33: runtime error:
load of value 131, which is not a valid value for type 'bool'
    #0 0x7e756f371a93 in SEL_ARG::clone(RANGE_OPT_PARAM*, SEL_ARG*, SEL_ARG**)

(cherry picked from commit 6be046a6b3e6c6cb43a9d1cca726e9e601294d50)

+++ unittest/gunit/mysys_my_time-t.cc
@@ -471,7 +471,8 @@ TEST(MysysMyTime, TimeZoneDisplacementToSeconds) {
   // Various invalid hours.
   EXPECT_TRUE(CheckTimeZoneDisplacement("+15"));
   EXPECT_TRUE(CheckTimeZoneDisplacement("-15"));
-  EXPECT_TRUE(CheckTimeZoneDisplacement(("+" + std::to_string(int_max + 1))))
+  EXPECT_TRUE(CheckTimeZoneDisplacement(
+      ("+" + std::to_string(static_cast<unsigned>(int_max) + 1))))
       << "Int wrap-around in hours not caught.";
 
   // Various invalid minutes.
@@ -480,7 +481,8 @@ TEST(MysysMyTime, TimeZoneDisplacementToSeconds) {
   EXPECT_TRUE(CheckTimeZoneDisplacement("+0:60"));
   EXPECT_TRUE(CheckTimeZoneDisplacement("+00:60"));
   EXPECT_TRUE(CheckTimeZoneDisplacement("+00:61"));
-  EXPECT_TRUE(CheckTimeZoneDisplacement(("+0:" + std::to_string(int_max + 1))))
+  EXPECT_TRUE(CheckTimeZoneDisplacement(
+      ("+0:" + std::to_string(static_cast<unsigned>(int_max) + 1))))
       << "Int wrap-around in minutes not caught.";
 
   EXPECT_TRUE(CheckTimeZoneDisplacement("-00:00"));
